apps/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
boto3==1.26.79
botocore==1.29.79
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
ftfy==6.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
mailchecker==5.0.7
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.0
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.6
platformdirs==2.6.2
pymongo==4.3.3
python-benedict==0.28.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-fsutil==0.10.0
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.8.2
